changed the html to fit the css and functionality I planned for.
Once I saw how my instructor was setting up the project, I decided to rethink what I wanted mine to look like. I now have a select with a few cities to choose from and a container that will scroll across the top the city and date like a marquee and the time will float largely in the middle. I will be working on how I want to incorporate the user's current time. I may have it scroll across the top of the page, fixed on top as if it were a nav bar