updated the scrolling cities
made an array instead of collecting the 491 cities I was first trying to gather. The scroll was only showing a handful, and so I gathered the information for all major cities, leaving out the three major ones that I have in the drop-down select element.

Hope you like my project! :)